backend/requirements.txt
# Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database (PostgreSQL + SQLAlchemy)
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0
cryptography==41.0.7

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
openpyxl==3.1.2
python-dateutil==2.8.2
chardet==5.2.0

# Financial Data APIs
yfinance==0.2.32
alpha-vantage==2.3.1

# Background Tasks & Caching
celery==5.3.4
redis==5.0.1

# WebSocket & Real-time
python-socketio==5.10.0

# AI/LLM Integration
ollama==0.1.6
langchain==0.1.0
langchain-community==0.0.13

# ML/NLP for Advanced Features
transformers>=4.36.0
torch>=2.1.0
scikit-learn>=1.3.2
tensorflow>=2.15.0
nltk>=3.8.1
textblob>=0.17.1
scipy>=1.11.4
sentencepiece>=0.1.99
accelerate>=0.25.0

# HTTP & Network
requests==2.31.0
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
pytz==2023.3
python-slugify==8.0.1
